requirements.txt
# Core dependencies
python-dotenv>=0.19.0
snowflake-connector-python>=2.7.0
tabulate>=0.8.9

# Schema management dependencies
alembic>=1.8.0
sqlalchemy>=1.4.0
snowflake-sqlalchemy>=1.7.3

# Additional utilities
pandas>=1.3.0
pyarrow>=7.0.0  # Often needed with Snowflake for optimized data transfer 